+++ Changelog.md
@@ -30,8 +30,8 @@ and this project adheres to
 
 - Docker Compose v1 (`docker-compose`) support. Only Docker Compose v2
   (`docker compose`) is supported now.
-- Drop support for Magento 2.0.x, 2.1.x, 2.2.x. These versions are not
-  installable anymore.
+- Drop support for Magento 2.0.x, 2.1.x, 2.2.x, 2.3.x, 2.4.0, 2.4.1 These
+  versions are not installable anymore.
 - Drop `magento:test-dockerfiles` comamnd.
 - Drop `bitnami/mariadb` images support.